Add files via upload
The Secret Message Encryption and Decryption tool is a Python GUI application built with Tkinter. This project allows users to encrypt their messages using a chosen encryption algorithm and a secret key. The encrypted message can then be shared securely with the intended recipient. The recipient can use the same tool to decrypt the message using the provided key, revealing the original content. This user-friendly application provides a convenient and secure way to exchange sensitive information while ensuring confidentiality.